+++ include/components/renderer_interface.hpp
@@ -10,7 +10,7 @@ class renderer_interface {
  public:
   renderer_interface(const tags::action_context& action_ctxt) : m_action_ctxt(action_ctxt){};
 
-  virtual void render_offset(const tags::context& ctxt, int pixels) = 0;
+  virtual void render_offset(const tags::context& ctxt, const extent_val offset) = 0;
   virtual void render_text(const tags::context& ctxt, const string&& str) = 0;
   virtual void change_alignment(const tags::context& ctxt) = 0;